+++ lib/outliers/verifications/shared.rb
@@ -7,8 +7,8 @@ module Outliers
       end
 
       def equals?(args)
-        keys = Array(args[:keys])
-        logger.debug "Verifying '#{list.join(',')}' equals '#{list.empty? ? 'no resources' : list_by_key.join(',')}'."
+        keys = Array(args)
+        logger.debug "Verifying '#{keys.join(',')}' equals '#{list.empty? ? 'no resources' : list_by_key.join(',')}'."
         list.reject {|r| keys.include? r.id}
       end